requirements.txt
# ═══════════════════════════════════════════════
# Expert Chatbot - Requirements
# ═══════════════════════════════════════════════

# ─── Web Framework ───────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.25.0

# ─── Streamlit Frontend ───────────────────────
streamlit>=1.31.0

# ─── Gemini AI ───────────────────────────────
google-generativeai>=0.5.0

# ─── Configuration & Validation ──────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ─── Rate Limiting ───────────────────────────
slowapi>=0.1.9

# ─── Retry / Resilience ───────────────────────
tenacity>=8.2.0

# ─── File Processing ─────────────────────────
Pillow>=10.2.0
pandas>=2.1.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# ─── Utilities ───────────────────────────────
numpy>=1.26.0
cachetools>=5.3.0
validators>=0.22.0
python-jose[cryptography]>=3.3.0

# ─── Logging ─────────────────────────────────
structlog>=24.1.0

# ─── Dev / Docs ──────────────────────────────
requests>=2.31.0